requirements.txt
flask
gunicorn
flask-mysql
flask-bcrypt
cryptography
python-dotenv

